Dev/reuseable table to development...

+++ RMS/src/Components/Table/Columns/CreateColumns.tsx
@@ -32,7 +32,7 @@ const createColumns = (baseColumns: { accessorKey: string; header: string; }[], 
             if (col.accessorKey === 'filekey' && rowIdAccessor === "reportid") {
                 const filekey = info.getValue();
                 const fileName = filekey.split('/').pop()?.split('-').pop();
-                const handleDownload = async (e:any) => {
+                const handleDownload = async (e: any) => {
                     e.preventDefault();
                     try {
                         const response = await axios.get(
@@ -40,7 +40,7 @@ const createColumns = (baseColumns: { accessorKey: string; header: string; }[], 
                             {
                                 responseType: "blob",
                                 headers: {
-                                    Authorization: `Bearer ${token}`, 
+                                    Authorization: `Bearer ${token}`,
                                 },
                             }
                         );
@@ -66,6 +66,42 @@ const createColumns = (baseColumns: { accessorKey: string; header: string; }[], 
                     </a>
                 );
             }
+            // if (col.accessorKey === 'status' && rowIdAccessor === "reportstatushistoryid") {
+            //     const status = info.getValue();
+            //     const handleDownload = async (e: any) => {
+            //         e.preventDefault();
+            //         try {
+            //             const response = await axios.get(
+            //                 `${import.meta.env.VITE_BACKEND_URL}/api/reports/download/${info.row.original[rowIdAccessor]}`,
+            //                 {
+            //                     responseType: "blob",
+            //                     headers: {
+            //                         Authorization: `Bearer ${token}`,
+            //                     },
+            //                 }
+            //             );
+
+            //             // Create a URL for the downloaded file
+            //             const url = window.URL.createObjectURL(new Blob([response.data]));
+            //             const link = document.createElement("a");
+            //             link.href = url;
+            //             link.setAttribute("download", fileName || "downloaded_file.xsl");
+            //             document.body.appendChild(link);
+            //             link.click();
+            //             document.body.removeChild(link);
+            //         } catch (error) {
+            //             console.error("Error downloading file", error);
+            //         }
+            //     };
+
+            //     return (
+            //         <a style={{ color: "#bc1a1a", fontWeight: "bold", textDecoration: "none", display: "flex", justifyContent: "center" }} href={`${import.meta.env.VITE_BACKEND_URL}/api/reports/download/${info.row.original[rowIdAccessor]}`} download onClick={handleDownload}>
+            //             {`${fileName}  `}
+            //             <CloudDownloadIcon sx={{ marginLeft: "0.5rem", marginTop: "-0.2rem" }}>
+            //             </CloudDownloadIcon>
+            //         </a>
+            //     );
+            // }
 
             return info.getValue();
         }
@@ -77,21 +113,25 @@ const createColumns = (baseColumns: { accessorKey: string; header: string; }[], 
             cell: (info: any) => <StatusSelect value={info.getValue()} rowId={info.row.original[rowIdAccessor]} handleStatusChange={handleStatusChange} />,
         });
     }
-    columns.push({
-        accessorKey: "edit",
-        header: "Actions",
-        cell: ({ row }: any) => (
-            <span style={{ display: "flex", justifyContent: "center", alignItems: "center", marginTop: "-0.5rem" }}>
-                {includeEdit && <IconButton onClick={() => onEdit(row.original)} sx={{ ml: 2, width: "auto", height: "auto" }}>
-                    <Edit />
-                </IconButton>}
+    if (rowIdAccessor != "reportstatushistoryid") {
 
-                <DeleteIcon
-                    sx={{ color: "#7d0e0e", cursor: "pointer" }}
-                    onClick={() => handleDeleteClick(row.original[rowIdAccessor])} />
-            </span>
-        ),
-    });
+        columns.push({
+            accessorKey: "edit",
+            header: "Actions",
+            cell: ({ row }: any) => (
+                <span style={{ display: "flex", justifyContent: "center", alignItems: "center", marginTop: "-0.5rem" }}>
+                    {includeEdit && <IconButton onClick={() => onEdit(row.original)} sx={{ ml: 2, width: "auto", height: "auto" }}>
+                        <Edit />
+                    </IconButton>}
+
+                    <DeleteIcon
+                        sx={{ color: "#7d0e0e", cursor: "pointer" }}
+                        onClick={() => handleDeleteClick(row.original[rowIdAccessor])} />
+                </span>
+            ),
+        });
+    }
+
     return columns;
 };
